feat(connect-iframe): standalone core build
