+++ packages/connect-iframe/package.json
@@ -4,7 +4,9 @@
     "private": true,
     "scripts": {
         "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
-        "build": "rimraf build && TS_NODE_PROJECT=\"tsconfig.json\" yarn webpack --config ./webpack/prod.webpack.config.ts",
+        "build:iframe": "TS_NODE_PROJECT=\"tsconfig.json\" yarn webpack --config ./webpack/prod.webpack.config.ts",
+        "build:core-module": "TS_NODE_PROJECT=\"tsconfig.json\" yarn webpack --config ./webpack/core.webpack.config.ts",
+        "build": "rimraf build && yarn build:iframe",
         "___NOTE__": "iframe build is one of the prerequisites of suite-web. build:lib script provides it together with other libraries",
         "build:lib": "yarn build",
         "type-check": "tsc --build tsconfig.json",
@@ -26,6 +28,7 @@
         "typescript": "4.9.5",
         "webpack": "^5.89.0",
         "webpack-cli": "^5.1.4",
+        "webpack-merge": "^5.10.0",
         "worker-loader": "^3.0.8"
     },
     "nx": {
