+++ packages/connect-web/tsconfig.json
@@ -12,6 +12,7 @@
     },
     "include": [
         ".",
+        "../connect-iframe/webpack/base.webpack.config.ts",
         "../connect-iframe/webpack/prod.webpack.config.ts",
         "../connect-iframe/package.json",
         "../connect-popup/webpack/prod.webpack.config.ts",
